more idempotency

+++ build/controller_all.sh
@@ -29,11 +29,13 @@ if [ "%%DEPLOY_SWIFT%%" = "yes" ]; then
   vgcreate swift /dev/xvdf1 || true
 
   for DISK in disk1 disk2 disk3; do
-    lvcreate -L 10G -n ${DISK} swift
-    echo "/dev/swift/${DISK} /srv/${DISK} xfs loop,noatime,nodiratime,nobarrier,logbufs=8 0 0" >> /etc/fstab
-    mkfs.xfs -f /dev/swift/${DISK}
-    mkdir -p /srv/${DISK}
-    mount /srv/${DISK}
+    if [ ! mountpoint /srv/${DISK} ]; then
+        lvcreate -L 10G -n ${DISK} swift
+        echo "/dev/swift/${DISK} /srv/${DISK} xfs loop,noatime,nodiratime,nobarrier,logbufs=8 0 0" >> /etc/fstab
+        mkfs.xfs -f /dev/swift/${DISK}
+        mkdir -p /srv/${DISK}
+        mount /srv/${DISK}
+    fi
   done
 fi
